Move the URL.checkFileURLExists() and URL.checkSubpathExists() utilities to their own appropriately located source file.
